+++ docutils/test/test_readers/test_pep/test_inline_markup.py
@@ -57,13 +57,13 @@ and RFC 2822 (which obsoletes RFC822 and RFC-733).
             pep-0287.txt
         ),
         and \n\
-        <reference refuri="https://tools.ietf.org/html/rfc2822.html">
+        <reference refuri="https://www.rfc-editor.org/info/rfc2822/">
             RFC 2822
          (which obsoletes \n\
-        <reference refuri="https://tools.ietf.org/html/rfc822.html">
+        <reference refuri="https://www.rfc-editor.org/info/rfc822/">
             RFC822
          and \n\
-        <reference refuri="https://tools.ietf.org/html/rfc733.html">
+        <reference refuri="https://www.rfc-editor.org/info/rfc733/">
             RFC-733
         ).
 """],
@@ -85,7 +85,7 @@ RFC
             PEP
             287
     <paragraph>
-        <reference refuri="https://tools.ietf.org/html/rfc2822.html">
+        <reference refuri="https://www.rfc-editor.org/info/rfc2822/">
             RFC
             2822
 """],
@@ -105,7 +105,7 @@ PEP 287 (https://peps.python.org/pep-0287), RFC 2822.
         <reference refuri="https://peps.python.org/pep-0287">
             https://peps.python.org/pep-0287
         ), \n\
-        <reference refuri="https://tools.ietf.org/html/rfc2822.html">
+        <reference refuri="https://www.rfc-editor.org/info/rfc2822/">
             RFC 2822
         .
 """],